plugins/rgb/display_image/requirements.txt
pillow==11.2.1
pluggy==1.6.0